[kernel] Add x86_64 (64-bit) support and CI coverage

+++ src/kernel/src/hal/arch/x86_64/cpu/exception/info.rs
@@ -56,8 +56,9 @@ impl ExceptionInformation {
         self.num as u32
     }
 
-    pub fn code(&self) -> u64 {
-        self.code
+    #[allow(clippy::as_conversions)]
+    pub fn code(&self) -> u32 {
+        self.code as u32
     }
 
     pub fn addr(&self) -> u64 {
